Fixing up configs
